Add compaction cost calculation Add legality check - currently only check the latency
in generateNewState - use Random that passed by SimulatedAnnealing class
previously this function create new randomGen obj everytime it was called.
This caused problem because srand value will be the same because time(NULL)
increments every second. But this function is called many times in one second
thus the result of rand() would be the same for every function call in that second.
This isn't what I want. So, instead of create new randomGen obj, I used the old one
from SA obj.

Todo:
- need to implement bandwidth legality check
- in generateNewState - need to check whether new position has a core or not